requirements.txt
streamlit
pypdf
pandas
google-generativeai
